+++ ports/msp430/qk/ccs/qfn_port.h
@@ -49,9 +49,9 @@
 /* QK-nano interrupt entry/exit */
 #define QK_ISR_ENTRY()   ((void)0)
 #define QK_ISR_EXIT()    do { \
-    if (QK_sched_() != (uint_fast8_t)0) { \
-        QK_activate_(); \
-    } \
+    if (QK_sched_() != 0U) {  \
+        QK_activate_();       \
+    }                         \
 } while (false)
